+++ symbol_table/symbol.c
@@ -3,7 +3,7 @@
 //
 
 #include "symbol.h"
-#include "../../config/config.h"
+#include "../config.h"
 
 
 #include <stdlib.h>
