+++ symbol_table/symbol_table.h
@@ -5,9 +5,9 @@
 #ifndef ASSEMBLER_SYMBOL_TABLE_H
 #define ASSEMBLER_SYMBOL_TABLE_H
 
-#include "../../config/config.h"
-#include "../symbol/symbol.h"
-#include "../symbol_node/symbol_node.h"
+#include "../config.h"
+#include "symbol.h"
+#include "symbol_node.h"
 
 #include <stdio.h>
 #include <stdbool.h>
